refactor(be): move study-room constants to libs
