+++ apps/backend/apps/client/src/study/interface/study-socket.interface.ts
@@ -1,21 +1,3 @@
-// Redis Key
-export const STUDY_KEY_PREFIX = 'study'
-export const STUDY_ROOM_PREFIX = `${STUDY_KEY_PREFIX}:room`
-export const STUDY_CHAT_PREFIX = `${STUDY_KEY_PREFIX}:chat`
-
-export const roomKey = (groupId: number) => `${STUDY_ROOM_PREFIX}:${groupId}`
-export const membersKey = (groupId: number) =>
-  `${STUDY_ROOM_PREFIX}:${groupId}:members`
-export const reconnectKey = (groupId: number, userId: number) =>
-  `${STUDY_ROOM_PREFIX}:${groupId}:reconnect:${userId}`
-
-export const chatKey = (groupId: number) => `${STUDY_CHAT_PREFIX}:${groupId}`
-
-// 상수
-export const RECONNECT_GRACE_SEC = 30 // 30초
-export const REMINDER_BEFORE_END_MS = 10 * 60 * 1000 // 10분
-export const BUFFER_SEC = 10 // 10초
-
 // Room
 export interface RoomState {
   endAt: number
