+++ apps/backend/libs/constants/src/index.ts
@@ -5,3 +5,4 @@ export * from './check.constants'
 export * from './submission.constants'
 export * from './argon2.constants'
 export * from './score-calculation.constants'
+export * from './study-room.constants'
